# server/docs/swagger.yaml
basePath: /
definitions:
...
  application.App:
    properties:
      createdAt:
        description: 创建时间
        type: string
      id:
        description: 主键ID
        type: integer
      name:
        description: 应用名称
        type: string
      type:
        description: 应用类型 0 未定义 1 数据库 2 缓存 3 应用 4 存储 5 负载均衡 6 备份 7 反向代理
        type: integer
      updatedAt:
        description: 更新时间
        type: string
      version:
        description: 版本
        type: string
    type: object
  application.ApplicationServer:
    properties:
      appIds:
        description: 安装应用列表
        type: string
      apps:
        description: 安装应用列表
        items:
          $ref: '#/definitions/application.App'
        type: array
      architecture:
        description: 架构
        type: integer
      createdAt:
        description: 创建时间
        type: string
      hostname:
        description: 主机名
        type: string
      id:
        description: 主键ID
        type: integer
      manageIp:
        description: 管理ip
        type: string
      os:
        description: 系统
        type: integer
      osVersion:
        description: 系统版本
        type: string
      sshPort:
        description: ssh端口
        type: integer
      systemId:
        description: 所属系统id
        type: integer
      updatedAt:
        description: 更新时间
        type: string
    type: object
  application.ApplicationSystem:
    properties:
      createdAt:
        description: 创建时间
        type: string
      id:
        description: 主键ID
        type: integer
      name:
        description: 系统名
        type: string
      position:
        description: 系统位置
        type: integer
      updatedAt:
        description: 更新时间
        type: string
    type: object
...
  application.SystemRelation:
    properties:
      createdAt:
        description: 创建时间
        type: string
      endSystem:
        $ref: '#/definitions/application.ApplicationSystem'
      endSystemId:
        description: 目的节点id',
        type: integer
      endSystemUrl:
        description: 目的节点url',
        type: string
      id:
        description: 主键ID
        type: integer
      relation:
        description: 调用关系',
        type: string
      startSystem:
        $ref: '#/definitions/application.ApplicationSystem'
      startSystemId:
        description: 源节点id
        type: integer
      updatedAt:
        description: 更新时间
        type: string
    type: object
...
  request.ExcelInfo:
    properties:
      fileName:
        description: 文件名
        type: string
      header:
        items:
          type: string
        type: array
      infoList:
        items:
          $ref: '#/definitions/application.ApplicationServer'
        type: array
    type: object
...
  request.GetById:
    properties:
      id:
        description: 主键ID
        type: number
      page:
        description: 页码
        type: integer
      pageSize:
        description: 每页大小
        type: integer
      sortBy:
        type: string
      sortInverted:
        type: boolean
    type: object
...
  request.IdsReq:
    properties:
      ids:
        items:
          type: integer
        type: array
    type: object
...
  request.PageInfo:
    properties:
      page:
        description: 页码
        type: integer
      pageSize:
        description: 每页大小
        type: integer
    type: object
...
info:
  contact: {}
  description: This is a sample Server pets
  title: Swagger Example API
  version: 0.0.1
paths:
...
  /cmdb/addApp:
    post:
      consumes:
      - application/json
      parameters:
      - description: 类型, 名称, 版本
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/application.App'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"添加成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 新增应用
      tags:
      - CmdbServer
  /cmdb/addServer:
    post:
      consumes:
      - application/json
      parameters:
      - description: 主机名, 架构, 管理ip, 系统, 系统版本
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/application.ApplicationServer'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"添加成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 新增服务器
      tags:
      - CmdbServer
...
  /cmdb/deleteApp:
    post:
      consumes:
      - application/json
      parameters:
      - description: 应用id
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.GetById'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"删除成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 删除应用
      tags:
      - CmdbServer
  /cmdb/deleteServer:
    post:
      consumes:
      - application/json
      parameters:
      - description: 服务器id
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.GetById'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"删除成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 删除服务器
      tags:
      - CmdbServer
  /cmdb/deleteServerByIds:
    post:
      consumes:
      - application/json
      parameters:
      - description: ID
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.IdsReq'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"删除成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 批量删除服务器
      tags:
      - CmdbServer
...
  /cmdb/downloadTemplate:
    get:
      consumes:
      - multipart/form-data
      produces:
      - application/json
      responses:
        "200":
          description: ""
      security:
      - ApiKeyAuth: []
      summary: 下载模板
      tags:
      - CmdbServer
  /cmdb/exportExcel:
    post:
      consumes:
      - application/json
      parameters:
      - description: 导出Excel文件信息
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.ExcelInfo'
      produces:
      - application/octet-stream
      responses:
        "200":
          description: ""
      security:
      - ApiKeyAuth: []
      summary: 导出Excel
      tags:
      - CmdbServer
...
  /cmdb/getAllServerIds:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"获取成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 获取所有服务器
      tags:
      - CmdbServer
  /cmdb/getAppById:
    post:
      consumes:
      - application/json
      parameters:
      - description: 服务器id
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.GetById'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"获取成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 根据id获取应用
      tags:
      - CmdbServer
  /cmdb/getAppList:
    post:
      consumes:
      - application/json
      parameters:
      - description: 页码, 每页大小
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.PageInfo'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"获取成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 分页获取基础app列表
      tags:
      - CmdbServer
...
  /cmdb/getServerById:
    post:
      consumes:
      - application/json
      parameters:
      - description: 服务器id
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.GetById'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"获取成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 根据id获取服务器
      tags:
      - CmdbServer
  /cmdb/getServerList:
    post:
      consumes:
      - application/json
      parameters:
      - description: 页码, 每页大小
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.PageInfo'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"获取成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 分页获取基础server列表
      tags:
      - CmdbServer
...
  /cmdb/getSystemServers:
    post:
      consumes:
      - application/json
      parameters:
      - description: 系统id
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.GetById'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"获取成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 根据系统id获取服务器
      tags:
      - CmdbServer
  /cmdb/importExcel:
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: 导入Excel文件
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"导入成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 导入Excel文件
      tags:
      - CmdbServer
  /cmdb/server/addRelation:
    post:
      consumes:
      - application/json
      parameters:
      - description: ' '
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/application.SystemRelation'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"添加成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 新增联系
      tags:
      - CmdbServer
  /cmdb/server/relations:
    post:
      consumes:
      - application/json
      parameters:
      - description: 服务器id
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.GetById'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"获取成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 获取关系图
      tags:
      - CmdbServer
...
  /cmdb/updateApp:
    post:
      consumes:
      - application/json
      parameters:
      - description: 类型, 名称, 版本
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/application.App'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"更新成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 更新应用
      tags:
      - CmdbServer
  /cmdb/updateServer:
    post:
      consumes:
      - application/json
      parameters:
      - description: 主机名, 架构, 管理ip, 系统, 系统版本
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/application.ApplicationServer'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"更新成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 更新服务器
      tags:
      - CmdbServer
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: x-token
    type: apiKey
swagger: "2.0"